Refactoring, Adding Tests, Fixed Bugs in Article/Image rel, Secured a…

+++ src/main/java/cz/kavka/service/SchoolAchievementsServiceImpl.java
@@ -9,10 +9,12 @@ import cz.kavka.service.serviceinterface.SchoolAchievementsService;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 @Service
+@Transactional(readOnly = true)
 public class SchoolAchievementsServiceImpl implements SchoolAchievementsService {
 
     private final SchoolAchievementsRepository schoolAchievementsRepository;
@@ -21,28 +23,47 @@ public class SchoolAchievementsServiceImpl implements SchoolAchievementsService 
 
     private final SchoolYearRepository schoolYearRepository;
 
+
     @Autowired
-    public SchoolAchievementsServiceImpl(SchoolAchievementsRepository schoolAchievementsRepository, SchoolAchievementsMapper schoolAchievementsMapper, SchoolYearRepository schoolYearRepository) {
+    public SchoolAchievementsServiceImpl(
+            SchoolAchievementsRepository schoolAchievementsRepository,
+            SchoolAchievementsMapper schoolAchievementsMapper,
+            SchoolYearRepository schoolYearRepository) {
         this.schoolAchievementsMapper = schoolAchievementsMapper;
         this.schoolAchievementsRepository = schoolAchievementsRepository;
         this.schoolYearRepository = schoolYearRepository;
+
     }
 
 
+    @Transactional
     @Override
     public SchoolAchievementsDTO createAchievement(SchoolAchievementsDTO schoolAchievementsDTO) {
+        if (schoolAchievementsDTO.getSchoolYear() == null ||
+                schoolAchievementsDTO.getContent() == null ||
+                schoolAchievementsDTO.getTitle() == null) {
+            throw new IllegalArgumentException("Error: entity must not by null");
+        }
         SchoolAchievementsEntity entityToSave = schoolAchievementsMapper.toEntity(schoolAchievementsDTO);
         entityToSave.setSchoolYear(schoolYearRepository.getReferenceById(schoolAchievementsDTO.getSchoolYear().getId()));
-        schoolAchievementsRepository.save(entityToSave);
+        SchoolAchievementsEntity savedEntity = schoolAchievementsRepository.save(entityToSave);
 
-        return schoolAchievementsMapper.toDTO(entityToSave);
+        return schoolAchievementsMapper.toDTO(savedEntity);
     }
 
+    @Transactional
     @Override
     public SchoolAchievementsDTO updateAchievement(Long id, SchoolAchievementsDTO schoolAchievementsDTO) {
-        schoolAchievementsDTO.setId(id);
-        SchoolAchievementsEntity editedEntity = schoolAchievementsRepository.save(schoolAchievementsMapper.toEntity(schoolAchievementsDTO));
-        return schoolAchievementsMapper.toDTO(editedEntity);
+        if (id == null) {
+            throw new IllegalArgumentException("Error: Id doesn't exist");
+        }
+        if (schoolAchievementsRepository.existsById(id)) {
+            schoolAchievementsDTO.setId(id);
+            SchoolAchievementsEntity editedEntity = schoolAchievementsRepository.save(schoolAchievementsMapper.toEntity(schoolAchievementsDTO));
+            return schoolAchievementsMapper.toDTO(editedEntity);
+        } else {
+            throw new EntityNotFoundException("Error: Entity not found");
+        }
     }
 
     @Override
@@ -55,13 +76,30 @@ public class SchoolAchievementsServiceImpl implements SchoolAchievementsService 
         return schoolAchievementsRepository.findAll().stream().map(schoolAchievementsMapper::toDTO).toList();
     }
 
+    @Transactional
     @Override
     public SchoolAchievementsDTO deleteAchievement(Long id) {
-        SchoolAchievementsEntity schoolAchievementsEntity = schoolAchievementsRepository.findById(id).orElseThrow(EntityNotFoundException::new);
-        SchoolAchievementsDTO deletedAchievement = schoolAchievementsMapper.toDTO(schoolAchievementsEntity);
-        schoolAchievementsRepository.delete(schoolAchievementsEntity);
+        if (schoolAchievementsRepository.existsById(id)) {
 
-        return deletedAchievement;
+
+            SchoolAchievementsEntity schoolAchievementsEntity = schoolAchievementsRepository.findById(id).orElseThrow(EntityNotFoundException::new);
+            SchoolAchievementsDTO deletedAchievement = schoolAchievementsMapper.toDTO(schoolAchievementsEntity);
+            schoolAchievementsRepository.delete(schoolAchievementsEntity);
+
+            return deletedAchievement;
+        } else {
+            throw new EntityNotFoundException("Error: Entity not found");
+        }
+
+    }
+
+    @Override
+    public List<SchoolAchievementsDTO> getAchievementsByYear(Long yearId) {
+
+        return schoolAchievementsRepository.getAllAchievementsByYear(yearId)
+                .stream()
+                .map(schoolAchievementsMapper::toDTO)
+                .toList();
 
     }
 }
