+++ src/main/java/cz/kavka/service/SchoolManagementServiceImpl.java
@@ -8,10 +8,12 @@ import cz.kavka.service.serviceinterface.SchoolManagementService;
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 @Service
+@Transactional(readOnly = true)
 public class SchoolManagementServiceImpl implements SchoolManagementService {
 
     private final SchoolManagementMapper schoolManagementMapper;
@@ -24,22 +26,34 @@ public class SchoolManagementServiceImpl implements SchoolManagementService {
         this.schoolManagementMapper = schoolManagementMapper;
     }
 
+    @Transactional
     @Override
     public SchoolManagementDTO createMember(SchoolManagementDTO schoolManagementDTO) {
+        if (schoolManagementDTO == null) {
+            throw new IllegalArgumentException("Error: Entity must not by null");
+        }
         SchoolManagementEntity savedEntity = schoolManagementRepository.save(schoolManagementMapper.toEntity(schoolManagementDTO));
         return schoolManagementMapper.toDTO(savedEntity);
     }
 
+    @Transactional
     @Override
     public SchoolManagementDTO editMember(SchoolManagementDTO schoolManagementDTO, Long id) {
-        schoolManagementDTO.setId(id);
-        SchoolManagementEntity editedEntity = schoolManagementRepository.save(schoolManagementMapper.toEntity(schoolManagementDTO));
-        return schoolManagementMapper.toDTO(editedEntity);
+        if (id == null) {
+            throw new IllegalArgumentException("Error: Id must not by null");
+        }
+        if (schoolManagementRepository.existsById(id)) {
+            schoolManagementDTO.setId(id);
+            SchoolManagementEntity editedEntity = schoolManagementRepository.save(schoolManagementMapper.toEntity(schoolManagementDTO));
+            return schoolManagementMapper.toDTO(editedEntity);
+        } else {
+            throw new EntityNotFoundException("Error: Entity to edit not found");
+        }
     }
 
     @Override
     public SchoolManagementDTO getMember(Long id) {
-        return schoolManagementMapper.toDTO(schoolManagementRepository.getReferenceById(id));
+        return schoolManagementMapper.toDTO(schoolManagementRepository.findById(id).orElseThrow(EntityNotFoundException::new));
     }
 
     @Override
@@ -47,11 +61,16 @@ public class SchoolManagementServiceImpl implements SchoolManagementService {
         return schoolManagementRepository.findAll().stream().map(schoolManagementMapper::toDTO).toList();
     }
 
+    @Transactional
     @Override
     public SchoolManagementDTO deleteMember(Long id) {
-        SchoolManagementEntity entityToDelete = schoolManagementRepository.findById(id).orElseThrow(EntityNotFoundException::new);
-        schoolManagementRepository.delete(entityToDelete);
-        return schoolManagementMapper.toDTO(entityToDelete);
+        if (schoolManagementRepository.existsById(id)) {
+            SchoolManagementEntity entityToDelete = schoolManagementRepository.findById(id).orElseThrow(EntityNotFoundException::new);
+            schoolManagementRepository.delete(entityToDelete);
+            return schoolManagementMapper.toDTO(entityToDelete);
+        } else {
+            throw new EntityNotFoundException("Error: Entity to delete not found");
+        }
 
     }
 }
